Add Gun, Shoot, Removed Junk.
Add a personal gun class, that splits the Gun, Character, and shooting mechanics into 3 separate blueprints instead of the default method of doing it.

Fixed Shoot, previously it had equip Gun to be able to shoot and hold a weapon now I made it to be automatically equip at the hand, Will change this in the future.